server/requirements.txt
alembic==1.12.1
aniso8601==9.0.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
bcrypt==4.1.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
command-not-found==0.3
cryptography==41.0.5
dbus-python==1.2.18
decorator==5.1.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.1.3
executing==2.0.0
Faker==19.13.0
filelock==3.12.4
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
importlib-resources==6.1.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.16.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.2.4
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2023.10.20
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyOpenSSL==23.3.0
pyparsing==2.4.7
pyrsistent==0.20.0
pysqlite3==0.5.2
pytest==7.4.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.22
SQLAlchemy-serializer==1.4.1
stack-data==0.6.3
systemd-python==234
tomli==2.0.1
traitlets==5.10.1
ttystatus==0.38
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.6
virtualenv==20.24.5
wadllib==1.3.6
wcwidth==0.2.8
Werkzeug==2.3.8
zipp==1.0.0
